requirements.txt
flask
tinytuya
python-dotenv
psutil
influxdb-client
